refactor: move concrete conversion
